move all component files to `/components` dir
